Module deno::module_graph

Structs

BundleOptions

CheckOptions

EmitOptions

Graph

A dependency graph of modules, were the modules that have been inserted via the builder will be loaded into the graph. Also provides an interface to be able to manipulate and handle the graph.

GraphBuilder

A structure for building a dependency graph of modules.

Module

A logical representation of a module within a graph.

ResultInfo

A structure that provides information about a module graph result.

Stats

TranspileOptions

A structure which provides options when transpiling modules.

Enums

BundleType

GraphError

A group of errors that represent errors that can occur when interacting with a module graph.

TypeLib

Represents the “default” type library that should be used when type checking the code in the module graph. Note that a user provided config of "lib" would override this value.

TypeScriptReference

An enum which represents the parsed out values of references in source code.

Functions

parse_deno_types

Determine if a comment contains a @deno-types pragma and optionally return its value.

parse_ts_reference

Determine if a comment contains a triple slash reference and optionally return its kind and value.