Trait deno::specifier_handler::SpecifierHandler

pub trait SpecifierHandler: Sync + Send {
    fn fetch(
        &mut self, 
        specifier: ModuleSpecifier, 
        maybe_location: Option<Location>, 
        is_dynamic: bool
    ) -> FetchFuture;

    fn get_tsbuildinfo(
        &self, 
        specifier: &ModuleSpecifier
    ) -> Result<Option<String>, AnyError>;

    fn set_cache(
        &mut self, 
        specifier: &ModuleSpecifier, 
        emit: &Emit
    ) -> Result<(), AnyError>;

    fn set_types(
        &mut self, 
        specifier: &ModuleSpecifier, 
        types: String
    ) -> Result<(), AnyError>;

    fn set_tsbuildinfo(
        &mut self, 
        specifier: &ModuleSpecifier, 
        tsbuildinfo: String
    ) -> Result<(), AnyError>;

    fn set_deps(
        &mut self, 
        specifier: &ModuleSpecifier, 
        dependencies: DependencyMap
    ) -> Result<(), AnyError>;

    fn set_version(
        &mut self, 
        specifier: &ModuleSpecifier, 
        version: String
    ) -> Result<(), AnyError>;
}

Required methods

fn fetch(
&mut self,
specifier: ModuleSpecifier,
maybe_location: Option<Location>,
is_dynamic: bool
) -> FetchFuture

Instructs the handler to fetch a specifier or retrieve its value from the cache.

fn get_tsbuildinfo(
&self,
specifier: &ModuleSpecifier
) -> Result<Option<String>, AnyError>

Get the optional build info from the cache for a given module specifier. Because build infos are only associated with the “root” modules, they are not expected to be cached for each module, but are “lazily” checked when a root module is identified. The emit_type also indicates what form of the module the build info is valid for.

fn set_cache(
&mut self,
specifier: &ModuleSpecifier,
emit: &Emit
) -> Result<(), AnyError>

Set the emit for the module specifier.

fn set_types(
&mut self,
specifier: &ModuleSpecifier,
types: String
) -> Result<(), AnyError>

When parsed out of a JavaScript module source, the triple slash reference to the types should be stored in the cache.

fn set_tsbuildinfo(
&mut self,
specifier: &ModuleSpecifier,
tsbuildinfo: String
) -> Result<(), AnyError>

Set the build info for a module specifier, also providing the cache type.

fn set_deps(
&mut self,
specifier: &ModuleSpecifier,
dependencies: DependencyMap
) -> Result<(), AnyError>

Set the graph dependencies for a given module specifier.

fn set_version(
&mut self,
specifier: &ModuleSpecifier,
version: String
) -> Result<(), AnyError>

Set the version of the source for a given module, which is used to help determine if a module needs to be re-emitted.