Module deno::tsc

Structs

EmittedFile

Request

A structure representing a request to be sent to the tsc runtime.

ResolveArgs

Response

State

Statics

COMPILER_SNAPSHOT

DENO_BROADCAST_CHANNEL_LIB

DENO_CONSOLE_LIB

DENO_CRYPTO_LIB

DENO_FETCH_LIB

DENO_FILE_LIB

DENO_NS_LIB

DENO_URL_LIB

DENO_WEBSOCKET_LIB

DENO_WEB_LIB

SHARED_GLOBALS_LIB

UNSTABLE_NS_LIB

WINDOW_LIB

Functions

compiler_snapshot

exec

Execute a request on the supplied snapshot, returning a response which contains information, like any emitted files, diagnostics, statistics and optionally an updated TypeScript build info.

get_asset

Retrieve a static asset that are included in the binary.

op