Struct deno::module_graph::Graph

pub struct Graph { /* fields omitted */ }

A dependency graph of modules, were the modules that have been inserted via the builder will be loaded into the graph. Also provides an interface to be able to manipulate and handle the graph.

Implementations

impl Graph

pub fn new(
handler: Arc<Mutex<dyn SpecifierHandler>>,
maybe_lockfile: Option<Arc<Mutex<Lockfile>>>
) -> Self

Create a new instance of a graph, ready to have modules loaded it.

The argument handler is an instance of a structure that implements the SpecifierHandler trait.

pub fn bundle(
&self,
options: BundleOptions
) -> Result<(String, Stats, Option<IgnoredCompilerOptions>), AnyError>

Transform the module graph into a single JavaScript module which is returned as a String in the result.

pub fn check(self, options: CheckOptions) -> Result<ResultInfo, AnyError>

Type check the module graph, corresponding to the options provided.

pub fn emit(
self,
options: EmitOptions
) -> Result<(HashMap<String, String>, ResultInfo), AnyError>

Emit the module graph in a specific format. This is specifically designed to be an “all-in-one” API for access by the runtime, allowing both emitting single modules as well as bundles, using Deno module resolution or supplied sources.

pub fn get_loadable_modules(
&self
) -> HashMap<ModuleSpecifier, Result<ModuleSource, AnyError>>

Retrieve a map that contains a representation of each module in the graph which can be used to provide code to a module loader without holding all the state to be able to operate on the graph.

pub fn get_media_type(&self, specifier: &ModuleSpecifier) -> Option<MediaType>

pub fn get_specifier(
&self,
specifier: &ModuleSpecifier
) -> Result<&Module, AnyError>

pub fn get_modules(&self) -> Vec<ModuleSpecifier>

Consume graph and return list of all module specifiers contained in the graph.

pub fn get_source(&self, specifier: &ModuleSpecifier) -> Option<String>

Get the source for a given module specifier. If the module is not part of the graph, the result will be None.

pub fn info(&self) -> Result<ModuleGraphInfo, AnyError>

Return a structure which provides information about the module graph and the relationship of the modules in the graph. This structure is used to provide information for the info subcommand.

pub fn lock(&self)

Verify the subresource integrity of the graph based upon the optional lockfile, updating the lockfile with any missing resources. This will error if any of the resources do not match their lock status.

pub fn resolve(
&self,
specifier: &str,
referrer: &ModuleSpecifier,
prefer_types: bool
) -> Result<ModuleSpecifier, AnyError>

Given a string specifier and a referring module specifier, provide the resulting module specifier and media type for the module that is part of the graph.

Arguments

specifier - The string form of the module specifier that needs to be resolved.

referrer - The referring ModuleSpecifier.

prefer_types - When resolving to a module specifier, determine if a type dependency is preferred over a code dependency. This is set to true when resolving module names for tsc as it needs the type dependency over the code, while other consumers do not handle type only dependencies.

pub fn transpile(
&mut self,
options: TranspileOptions
) -> Result<ResultInfo, AnyError>

Transpile (only transform) the graph, updating any emitted modules with the specifier handler. The result contains any performance stats from the compiler and optionally any user provided configuration compiler options that were ignored.

Arguments

options - A structure of options which impact how the code is transpiled.